add_firstStage

+++ working/src/main/java/ua/chernonog/working/antifraud/model/request/TransactionReq.java
@@ -6,5 +6,5 @@ import lombok.*;
 @Getter@Setter
 @NoArgsConstructor
 public class TransactionReq {
-    int amount;
+    long amount;
 }
